refactor stimulus_reflex.js (#444)
* removed local variables from stimulus_reflex.js

* callbacks extracted, CableReady processing extracted

* consolidate getConsumer into action_cable.js, merge extendStimulusController into register

* fixed reflexRoot assignment bug
