+++ javascript/controllers.js
@@ -1,4 +1,5 @@
 import { attributeValues } from './attributes'
+import { extractReflexName } from './utils'
 
 // Returns StimulusReflex controllers local to the passed element based on the data-controller attribute.
 //
@@ -22,4 +23,21 @@ export const allReflexControllers = (app, element) => {
     element = element.parentElement
   }
   return controllers
+}
+
+// Given a reflex string such as 'click->TestReflex#create' and a list of
+// controllers. It will find the matching controller based on the controller's
+// identifier. e.g. Given these controller identifiers ['foo', 'bar', 'test'],
+// it would select the 'test' controller.
+export const findControllerByReflexName = (reflexName, controllers) => {
+  const controller = controllers.find(controller => {
+    if (!controller.identifier) return
+
+    return (
+      extractReflexName(reflexName).toLowerCase() ===
+      controller.identifier.toLowerCase()
+    )
+  })
+
+  return controller || controllers[0]
 }
