+++ javascript/isolation_mode.js
@@ -1,0 +1,10 @@
+let isolationMode = false
+
+export default {
+  get disabled () {
+    return !isolationMode
+  },
+  set (value) {
+    isolationMode = value
+  }
+}
