+++ javascript/log.js
@@ -1,3 +1,5 @@
+import reflexes from './reflexes'
+
 function request (
   reflexId,
   target,
